scripts/solis_s6_app/requirements.txt
# Solis S6 11.4kW inverter control UI - run on 10.10.53.92
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymodbus>=3.6.0
jinja2>=3.1.0
python-multipart>=0.0.6
paho-mqtt>=2.0.0
